requirements.txt
streamlit
pandas
openpyxl
camelot-py[cv]